add relations to users and ideas

+++ src/idea/idea.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { IdeaEntity } from "./entities/idea.entity";
 import { Repository } from "typeorm";
 import { CreateIdeaDto } from "./dto/createIdea.dto";
 import { UpdateIdeaDto } from "./dto/updateIdea.dto";
+import { UserEntity } from "src/user/entities/user.entity";
 
 
 @Injectable()
@@ -11,16 +12,27 @@ export class IdeaService {
 
     constructor (
     @InjectRepository(IdeaEntity) 
-    private ideaRepository:Repository<IdeaEntity>
+    private ideaRepository:Repository<IdeaEntity>,
+    @InjectRepository(UserEntity) 
+    private userRepository:Repository<UserEntity>
     ){}
-   async getIdeas(){
-        return await this.ideaRepository.find()
+
+   async getIdeas(id?:number){
+        let where = {}
+        if(id)
+        where = {author:{id}}
+        return await this.ideaRepository.find({
+          where
+        })
     }
 
-    async createIdea(ideaDto:CreateIdeaDto){
-        console.log("ideaDto",ideaDto);
+    async createIdea(id:number,ideaDto:CreateIdeaDto){
         
+        const user = await this.userRepository.findOne({where:{id}})
         const idea = await this.ideaRepository.create(ideaDto)
+        idea.author = user
+        console.log("user",user);
+        
         await this.ideaRepository.save(idea)
         return idea;
     }
@@ -32,23 +44,33 @@ export class IdeaService {
       }
       return idea
     }
-    async updateSingleIdea(id:number,ideaDto:UpdateIdeaDto){
+
+    async updateSingleIdea(id:number,ideaDto:UpdateIdeaDto,userId:number){
       // await this.ideaRepository.update({id},ideaDto)
-      const idea = await this.ideaRepository.findOne({where:{id}})
+      const idea = await this.ideaRepository.findOne({   where:{id}} )
     //  Object.assign(idea,ideaDto)
       if(!idea){
         throw new NotFoundException("idea not found")
      }
+     this.ensureOwnership(idea,userId)
      await this.ideaRepository.update({id},ideaDto)
      //idea.save()
       return idea  
     }
-    async deleteSingleIdea(id:number){
+
+    async deleteSingleIdea(id:number,userId:number){
       const idea = await this.ideaRepository.findOne({where:{id}});
       if(!idea){
         throw new NotFoundException("idea not found")
      }
+     
+     this.ensureOwnership(idea,userId)
      await this.ideaRepository.delete({id});
-      return idea;  
+     return idea;  
+    }
+
+    private ensureOwnership(idea:IdeaEntity,ownerId:number){
+        if(idea.author.id !== ownerId)
+        throw new UnauthorizedException("This idea does not belong to you")
     }
 }
